refactor: extract Qwen-VL partial-prefix prefill into a mixin
qwen_vl.py had grown into a ~630-line god class. Move the partial
prefix-cache tail-reconstruction machinery (~156 lines) into a new
qwen_vl_prefill module as QwenVLPartialPrefixMixin, and unify the
per-modality (image/video) kwarg/config layout that four prefill methods
each hardcoded into a single iter_modalities helper. The main class
drops from 767 to 604 lines; behavior is unchanged (Qwen3-VL overrides
still resolve via MRO).
